feat(sparksql): Address review feedback for transform_values
- Move transform_values doc entry after size in map.rst (alphabetical order)
- Move TransformValuesTest to a dedicated TransformValuesTest.cpp
- Update nullValueInMap test to use v * 2 on {1:10, 2:null, 3:30},
  verifying that null values in the map produce null outputs
- Add blank lines between tests in TransformValuesTest.cpp
