+++ velox/functions/sparksql/tests/CMakeLists.txt
@@ -82,6 +82,7 @@ add_executable(
   ToJsonTest.cpp
   ToPrettyStringTest.cpp
   TransformTest.cpp
+  TransformValuesTest.cpp
   UnBase64Test.cpp
   UnscaledValueFunctionTest.cpp
   UpperLowerTest.cpp
